separates bars from main and created initial route
